Update index.html
This is great work, I can't believe how far ahead you are. I've just made one amendment, you were calling the   setInterval(startTime, 500); within the start time function itself. The problem with this is you're constructing what's know as a recursive call - 'a function that calls itself' . Every time the function is called it doubles the memory usage, this can crash the browser.